requirements.txt
torch
torchvision
timm
pandas
numpy
scikit-learn
matplotlib
seaborn
opencv-python
tqdm
wandb
shap
grad-cam
scikit-image
pillow